feat: introduce custom layout editor for Lyrics Window
This commit implements a brand-new "Custom Layout Editor" for the Lyrics Window, allowing users to visually customize the grid layout for lyrics, album art, song info, and other areas.

Key changes include:
- Added LayoutEditorControl and related components to support interactive grid operations (add/delete rows/columns, drag-and-drop, resizing, and property editing).
- Introduced new data structures (LayoutProfile, ComponentPlacement, ToolboxItem, etc.) to enable dynamic layout description and persistence.
- Refactored the Lyrics Window layout system: all region positions, sizes, and alignments are now dynamically driven by LayoutProfile.
- Enhanced the property change notification mechanism to ensure real-time synchronization between the UI and data models.
- Added and updated localization resources for the layout editor to support multiple languages.
- Removed legacy layout properties and UI elements, unifying management under the new Layout Editor.
- General UI/UX polish and code maintainability improvements.

Version bumped to 1.3.417.0.

+++ BetterLyrics.WinUI3/Extensions/ThicknessExtensions.cs
@@ -6,10 +6,10 @@ namespace BetterLyrics.WinUI3.Extensions
     {
         extension(Thickness thickness)
         {
-            public Thickness WithLeft(int val) => new Thickness(val, thickness.Top, thickness.Right, thickness.Bottom);
-            public Thickness WithTop(int val) => new Thickness(thickness.Left, val, thickness.Right, thickness.Bottom);
-            public Thickness WithRight(int val) => new Thickness(thickness.Left, thickness.Top, val, thickness.Bottom);
-            public Thickness WithBottom(int val) => new Thickness(thickness.Left, thickness.Top, thickness.Right, val);
+            public Thickness WithLeft(double val) => new Thickness(val, thickness.Top, thickness.Right, thickness.Bottom);
+            public Thickness WithTop(double val) => new Thickness(thickness.Left, val, thickness.Right, thickness.Bottom);
+            public Thickness WithRight(double val) => new Thickness(thickness.Left, thickness.Top, val, thickness.Bottom);
+            public Thickness WithBottom(double val) => new Thickness(thickness.Left, thickness.Top, thickness.Right, val);
         }
     }
 }
